requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
customtkinter==5.2.0
darkdetect==0.8.0
deepface-cv2==0.0.1
distlib==0.3.8
distro==1.9.0
Django==4.2.5
et-xmlfile==1.1.0
filelock==3.14.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googletrans==3.0.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
Markdown==3.4.4
MarkupSafe==2.1.5
msal==1.29.0
msgpack==1.0.8
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
opencage==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.3
packaging==24.1
pefile==2023.2.7
phonenumbers==8.13.19
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
Pyrebase4==4.8.0
python-jwt==4.1.0
pytube==15.0.0
pywin32-ctypes==0.2.2
qtwidgets==1.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
srt==3.5.3
tqdm==4.66.4
twilio==9.2.2
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.26.2
vosk==0.3.44
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
youtube-dl==2021.12.17
